Refactor and opimize shared prop types inference for pickers
