+++ lib/src/DatePicker/DatePicker.tsx
@@ -1,32 +1,22 @@
 import { useUtils } from '../_shared/hooks/useUtils';
-import { MaterialUiPickersDate } from '../typings/date';
 import { DatePickerToolbar } from './DatePickerToolbar';
 import { getFormatByViews } from '../_helpers/date-utils';
+import { WithViewsProps } from '../Picker/WithViewsProps';
 import { datePickerDefaultProps } from '../constants/prop-types';
 import { ResponsiveWrapper } from '../wrappers/ResponsiveWrapper';
 import { ExportedCalendarViewProps } from '../views/Calendar/CalendarView';
-import { WithViewsProps, WithDateInputProps } from "../Picker/WithViewsProps";
+import { makePickerWithStateAndWrapper } from '../Picker/makePickerWithState';
 import { ModalWrapper, InlineWrapper, StaticWrapper } from '../wrappers/Wrapper';
-import {
-  makePickerWithStateAndWrapper,
-} from '../Picker/makePickerWithState';
 
 export type DatePickerView = 'year' | 'date' | 'month';
 
-export interface BaseDatePickerProps extends ExportedCalendarViewProps {
-  /** Callback firing on year change @DateIOType */
-  onYearChange?: (date: MaterialUiPickersDate) => void;
-  /** Date format, that is displaying in toolbar */
-  toolbarFormat?: string;
-}
-
-export type DatePickerProps = BaseDatePickerProps &
-  WithDateInputProps &
-  WithViewsProps<'year' | 'date' | 'month'>;
+export interface BaseDatePickerProps
+  extends WithViewsProps<'year' | 'date' | 'month'>,
+    ExportedCalendarViewProps {}
 
 const datePickerConfig = {
   DefaultToolbarComponent: DatePickerToolbar,
-  useDefaultProps: ({ openTo = 'date', views = ['year', 'date'] }: DatePickerProps) => {
+  useDefaultProps: ({ openTo = 'date', views = ['year', 'date'] }: BaseDatePickerProps) => {
     const utils = useUtils();
 
     return {
@@ -39,22 +29,22 @@ const datePickerConfig = {
   },
 };
 
-export const DatePicker = makePickerWithStateAndWrapper<DatePickerProps>(
+export const DatePicker = makePickerWithStateAndWrapper<BaseDatePickerProps>(
   ResponsiveWrapper,
   datePickerConfig
 );
 
-export const MobileDatePicker = makePickerWithStateAndWrapper<DatePickerProps>(
+export const MobileDatePicker = makePickerWithStateAndWrapper<BaseDatePickerProps>(
   ModalWrapper,
   datePickerConfig
 );
 
-export const DesktopDatePicker = makePickerWithStateAndWrapper<DatePickerProps>(
+export const DesktopDatePicker = makePickerWithStateAndWrapper<BaseDatePickerProps>(
   InlineWrapper,
   datePickerConfig
 );
 
-export const StaticDatePicker = makePickerWithStateAndWrapper<DatePickerProps>(
+export const StaticDatePicker = makePickerWithStateAndWrapper<BaseDatePickerProps>(
   StaticWrapper,
   datePickerConfig
 );
