+++ lib/src/Picker/Picker.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import clsx from 'clsx';
-import { WithViewsProps } from './WithViewsProps';
 import { WrapperVariant } from '../wrappers/Wrapper';
 import { useViews } from '../_shared/hooks/useViews';
 import { makeStyles } from '@material-ui/core/styles';
@@ -9,42 +8,42 @@ import { ParsableDate } from '../constants/prop-types';
 import { BasePickerProps } from '../typings/BasePicker';
 import { MaterialUiPickersDate } from '../typings/date';
 import { DateInputProps } from '../_shared/PureDateInput';
-import { CalendarView } from '../views/Calendar/CalendarView';
+import { DatePickerView } from '../DatePicker/DatePicker';
+import { WithViewsProps, AnyPickerView } from './WithViewsProps';
 import { useIsLandscape } from '../_shared/hooks/useIsLandscape';
 import { DIALOG_WIDTH, VIEW_HEIGHT } from '../constants/dimensions';
 import { WrapperVariantContext } from '../wrappers/WrapperVariantContext';
 import { MobileKeyboardInputView } from '../views/MobileKeyboardInputView';
 import { ClockView, ExportedClockViewProps } from '../views/Clock/ClockView';
-import { BaseDatePickerProps, DatePickerView } from '../DatePicker/DatePicker';
+import { CalendarView, ExportedCalendarViewProps } from '../views/Calendar/CalendarView';
 
-export type PickerView = DateTimePickerView;
+type CalendarAndClockProps = ExportedCalendarViewProps & ExportedClockViewProps;
 
-export type ToolbarComponentProps<T extends PickerView = any> = BaseDatePickerProps &
-  ExportedClockViewProps & {
-    views: T[];
-    openView: T;
-    date: MaterialUiPickersDate;
-    setOpenView: (view: T) => void;
-    onChange: (date: MaterialUiPickersDate, isFinish?: boolean) => void;
-    toolbarTitle?: string;
-    // TODO move out, cause it is DateTimePickerOnly
-    hideTabs?: boolean;
-    dateRangeIcon?: React.ReactNode;
-    timeIcon?: React.ReactNode;
-    isLandscape: boolean;
-    ampmInClock?: boolean;
-    isMobileKeyboardViewOpen: boolean;
-    toggleMobileKeyboardView: () => void;
-    getMobileKeyboardInputViewButtonText?: () => string;
-  };
+export type ToolbarComponentProps<
+  T extends AnyPickerView = AnyPickerView
+> = CalendarAndClockProps & {
+  views: T[];
+  openView: T;
+  date: MaterialUiPickersDate;
+  setOpenView: (view: T) => void;
+  onChange: (date: MaterialUiPickersDate, isFinish?: boolean) => void;
+  toolbarTitle?: React.ReactNode;
+  toolbarFormat?: string;
+  // TODO move out, cause it is DateTimePickerOnly
+  hideTabs?: boolean;
+  dateRangeIcon?: React.ReactNode;
+  timeIcon?: React.ReactNode;
+  isLandscape: boolean;
+  ampmInClock?: boolean;
+  isMobileKeyboardViewOpen: boolean;
+  toggleMobileKeyboardView: () => void;
+  getMobileKeyboardInputViewButtonText?: () => string;
+};
 
-export interface ExportedPickerProps<TInputValue, TDateValue>
+export interface ExportedPickerProps<TView extends AnyPickerView>
   extends Omit<BasePickerProps, 'value' | 'onChange'>,
-    BaseDatePickerProps,
-    ExportedClockViewProps {
-  toolbarTitle?: string;
-  showToolbar?: boolean;
-  ToolbarComponent?: React.ComponentType<ToolbarComponentProps<any>>;
+    CalendarAndClockProps,
+    WithViewsProps<TView> {
   // TODO move out, cause it is DateTimePickerOnly
   hideTabs?: boolean;
   dateRangeIcon?: React.ReactNode;
@@ -52,10 +51,10 @@ export interface ExportedPickerProps<TInputValue, TDateValue>
 }
 
 export interface PickerProps<
-  T extends PickerView,
+  TView extends AnyPickerView,
   TInputValue = ParsableDate,
   TDateValue = MaterialUiPickersDate
-> extends ExportedPickerProps<TInputValue, TDateValue>, WithViewsProps<T> {
+> extends ExportedPickerProps<TView> {
   isMobileKeyboardViewOpen: boolean;
   toggleMobileKeyboardView: () => void;
   DateInputProps: DateInputProps<TInputValue, TDateValue>;
@@ -106,8 +105,9 @@ export function Picker({
   DateInputProps,
   isMobileKeyboardViewOpen,
   toggleMobileKeyboardView,
+  toolbarFormat,
   ...other
-}: PickerProps<PickerView>) {
+}: PickerProps<AnyPickerView>) {
   const classes = useStyles();
   const isLandscape = useIsLandscape(views, orientation);
   const wrapperVariant = React.useContext(WrapperVariantContext);
@@ -145,7 +145,7 @@ export function Picker({
           setOpenView={setOpenView}
           openView={openView}
           toolbarTitle={toolbarTitle}
-          ampmInClock={other.ampmInClock}
+          toolbarFormat={toolbarFormat}
           isMobileKeyboardViewOpen={isMobileKeyboardViewOpen}
           toggleMobileKeyboardView={toggleMobileKeyboardView}
         />
