+++ lib/src/_shared/hooks/useViews.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { PickerView } from '../../Picker/Picker';
 import { arrayIncludes } from '../../_helpers/utils';
 import { MaterialUiPickersDate } from '../../typings/date';
+import { AnyPickerView } from '../../Picker/WithViewsProps';
 
 export type PickerOnChangeFn = (date: MaterialUiPickersDate, isFinish?: boolean | symbol) => void;
 
@@ -12,8 +12,8 @@ export function useViews({
   isMobileKeyboardViewOpen,
   toggleMobileKeyboardView,
 }: {
-  views: PickerView[];
-  openTo: PickerView;
+  views: AnyPickerView[];
+  openTo: AnyPickerView;
   onChange: PickerOnChangeFn;
   isMobileKeyboardViewOpen: boolean;
   toggleMobileKeyboardView: () => void;
